MOved SHR Poatal to SHR app, made billing empty state ui consistent with other, same to error state.Fixed nav link issue of homepage billing and patient chart billing
