+++ packages/esm-billing-app/src/dashboard.meta.ts
@@ -5,12 +5,6 @@ export const dashboardMeta = {
   path: 'Billing',
 };
 
-export const shrSummaryDashboardMeta = {
-  slot: 'patient-chart-shr-summary-dashboard-slot',
-  columns: 1,
-  title: 'SHR Summary',
-  path: 'SHR Summary',
-};
 export const benefitsPackageDashboardMeta = {
   slot: 'patient-chart-benefits-dashboard-slot',
   columns: 2,
