--- packages/esm-billing-app/src/hooks/usePatientVisitSHRSummary.ts
@@ -1,9 +1,0 @@
-const usePatientVisitSHRSummary = () => {
-  return {
-    isLoading: false,
-    error: undefined,
-    shrData: undefined,
-  };
-};
-
-export default usePatientVisitSHRSummary;
