+++ packages/esm-billing-app/src/routes.json
@@ -19,7 +19,7 @@
     {
       "component": "billingDashboardLink",
       "name": "billing-dashboard-link",
-      "slot": "patient-chart-dashboard-slot",
+      "slot": "homepage-dashboard-slot",
       "meta": {
         "name": "billing",
         "title": "billing",
@@ -48,11 +48,7 @@
       "name": "billing-dashboard-root",
       "slot": "billing-dashboard-slot"
     },
-    {
-      "component": "shrAuthorizationForm",
-      "name": "shr-authorization-form"
-    }
-,    {
+   {
       "name": "billing-patient-summary",
       "component": "billingPatientSummary",
       "slot": "patient-chart-billing-dashboard-slot",
@@ -75,27 +71,13 @@
 
       }
     },
-    {
-      "name": "patient-chart-shr-summary-dashboard-link",
-      "component": "shrSummaryDashboardLink",
-      "slot": "patient-chart-dashboard-slot",
-      "meta": {
-        "columns": 1,
-        "columnSpan": 1,
-        "slot": "patient-chart-shr-summary-dashboard-slot",
-        "path": "SHR Summary"
-      }
-    },
+    
     {
       "name": "billing-check-in-form",
       "slot": "extra-visit-attribute-slot",
       "component": "billingCheckInForm"
     },
-    {
-      "name": "SHR Summary",
-      "slot": "patient-chart-shr-summary-dashboard-slot",
-      "component": "patientSHRSummary"
-    },
+    
     {
       "slot": "system-admin-page-card-link-slot",
       "component": "billableServicesCardLink",
