+++ packages/esm-billing-app/translations/en.json
@@ -35,6 +35,7 @@
   "claimJustification": "Claim Justification",
   "claimsItems": "Claims Items",
   "claimSummary": "Claim Summary",
+  "Conditopns": "Conditions",
   "createClaimError": "Create Claim error",
   "date": "Date of Claimed",
   "diagnoses": "Diagnoses",
@@ -135,6 +136,7 @@
   "shaNotValid": "Invalid SHA Number",
   "shaNumber": "SHA Number",
   "shortName": "Short Name",
+  "shrRecords": "SHR Records",
   "status": "Service Status",
   "stockItem": "Stock Item",
   "total": "Total",
