+++ packages/esm-shr-app/src/dashboard.meta.ts
@@ -13,3 +13,10 @@ export const referralDashboardMeta = {
   moduleName: '@kenyaemr/esm-shr-app',
   config: {},
 };
+
+export const shrSummaryDashboardMeta = {
+  slot: 'patient-chart-shr-summary-dashboard-slot',
+  columns: 1,
+  title: 'SHR Portal',
+  path: 'SHR Portal',
+};
