+++ packages/esm-shr-app/src/index.ts
@@ -1,12 +1,14 @@
 import { defineConfigSchema, getAsyncLifecycle, getSyncLifecycle, registerBreadcrumbs } from '@openmrs/esm-framework';
 import { configSchema } from './config-schema';
-import { referralDashboardMeta } from './dashboard.meta';
+import { referralDashboardMeta, shrSummaryDashboardMeta } from './dashboard.meta';
 import { createDashboardLink } from '@openmrs/esm-patient-common-lib';
 import shrPatientSummaryComponent from './shrpatient-summary/shrpatient-summary.component';
 import { createHomeDashboardLink } from './create-dashboard-link';
 import ReferralReasonsDialogPopup from './referrals/referral-reasons/referral-reasons.component';
 import ReferralChartView from './referrals/patient-chart/referral-chart-view.component';
 import FacilityRefferalForm from './workspace/referrals.workspace.component';
+import SHRAuthorizationForm from './shr-summary/shr-authorization-form.workspace';
+import SHRSummaryPanell from './shr-summary/shr-summary.component';
 
 const moduleName = '@kenyaemr/esm-shr-app';
 
@@ -35,6 +37,11 @@ export const ReferralsDashboardLink = getSyncLifecycle(
   options,
 );
 
+export const shrSummaryDashboardLink = getSyncLifecycle(
+  createDashboardLink({ ...shrSummaryDashboardMeta, moduleName }),
+  options,
+);
+
 export const shrRoot = getAsyncLifecycle(() => import('./shr-root.component'), options);
 
 export const referralReasonsDialogPopup = getSyncLifecycle(ReferralReasonsDialogPopup, {
@@ -46,3 +53,7 @@ export const referralReasonsDialogPopup = getSyncLifecycle(ReferralReasonsDialog
 export const referralWidget = getSyncLifecycle(ReferralChartView, options);
 export const referralLink = getSyncLifecycle(createDashboardLink(referralDashboardMeta), options);
 export const facilityRefferalForm = getSyncLifecycle(FacilityRefferalForm, options);
+
+// SHR Summary
+export const patientSHRSummary = getSyncLifecycle(SHRSummaryPanell, options);
+export const shrAuthorizationForm = getSyncLifecycle(SHRAuthorizationForm, options);
