+++ packages/esm-shr-app/src/shr-summary/shr-summary.scss
