+++ packages/esm-shr-app/src/shr-summary/tables/shr-summary-table.component.tsx
