+++ packages/esm-shr-app/src/shr-summary/tables/shr-tables.scss
